bot/requirements.txt
python-telegram-bot[job-queue]==21.6
supabase==2.7.4
flask==3.0.3
pillow==10.4.0
pandas==2.2.2
openpyxl==3.1.5
httpx==0.27.2
pytz==2024.1
ethiopian-date==1.2.0
